fixing endpoint paths

+++ Backend/MovieRecommendationSystem/src/main/java/pt/uc/movierecommendation/movierecommendationsystem/Service/HistoryService.java
@@ -49,7 +49,8 @@ public class HistoryService {
     @Transactional
     public boolean remove(long movieId, long userId) {
         Optional<HistoryItem> item = historyItemRepository.findByUser_IdAndMovie_Id(userId, movieId);
-        if (item == null) return false;
+        if (item == null || item.isEmpty()) return false;
+    
         historyItemRepository.delete(item.get());
         return true;
